Enable dynamic support for `libdecor` inside SDL3
Done by locally building & installing `libdecor` before building SDL3.

Note that the `libdecor` binaries aren't shipped here, because I can't
make them work inside C# apps. It looks like both SDL and `libdecor`
do some level of automagic lookup of shared objects; SDL does it to find
`libdecor`, and `libdecor` does it to find its "plugins" which are
really implementations of the window decoration (if it fails to find
any, it just disables itself).

So this is just going to depend on system-ambient `libdecor` to be
present, and if there is none, then it'll fail to work. See if I care.

At this point I would like to extend my thanks to the gnome team for
wasting a day of my life on all this.

Note that x86 SDL binaries will not have this `libdecor` support enabled,
because SDL uses cmake and `libdecor` uses meson, and there is no
portable way to tell them both to build a 32-bit binary. Well, that's a
half-truth; you can pass CFLAGS to meson via envvars, *but* the install
script later will just install the resulting binaries as if they were
x64, which leads SDL cmake to not find them, and therefore not enable
dynamic `libdecor` support. So let's just not bother.