Bump ubuntu images & runners to 22.04
While this may mean the binaries produced will no longer be compatible
with some unix installations with older glibc versions present,
GitHub are pulling support for the `ubuntu-20.04` runner on April 1,
so we don't really have much of a say in this matter.

The docker image tags could possibly stay for a while longer, but the
whole reason why GitHub are pulling the 20.04 runners is that the LTS
period for Ubuntu 20.04 *also* ends on April 1, so maybe best to just
get it out of the way all across the board.